Refactor updateFeet into GCPlayerHandler.  Fix warnings for unused imports.

+++ src/main/java/micdoodle8/mods/galacticraft/core/entities/EntityEvolvedSkeleton.java
@@ -10,8 +10,6 @@ import net.minecraft.entity.ai.*;
 import net.minecraft.entity.monster.EntitySkeleton;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.projectile.EntityArrow;
-import net.minecraft.init.Items;
-import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
 
 public class EntityEvolvedSkeleton extends EntitySkeleton implements IEntityBreathable
@@ -46,14 +44,14 @@ public class EntityEvolvedSkeleton extends EntitySkeleton implements IEntityBrea
 
     public void attackEntityWithRangedAttack(EntityLivingBase par1EntityLivingBase, float par2)
     {
-        EntityArrow entityarrow = new EntityArrow(this.worldObj, this, par1EntityLivingBase, 0.8F, (float)(50 - this.worldObj.difficultySetting.getDifficultyId() * 4));
+        EntityArrow entityarrow = new EntityArrow(this.worldObj, this, par1EntityLivingBase, 0.8F, 50 - this.worldObj.difficultySetting.getDifficultyId() * 4);
         int i = EnchantmentHelper.getEnchantmentLevel(Enchantment.power.effectId, this.getHeldItem());
         int j = EnchantmentHelper.getEnchantmentLevel(Enchantment.punch.effectId, this.getHeldItem());
-        entityarrow.setDamage((double)(par2 * 2.0F) + this.rand.nextGaussian() * 0.25D + (double)((float)this.worldObj.difficultySetting.getDifficultyId() * 0.11F));
+        entityarrow.setDamage(par2 * 2.0F + this.rand.nextGaussian() * 0.25D + this.worldObj.difficultySetting.getDifficultyId() * 0.11F);
 
         if (i > 0)
         {
-            entityarrow.setDamage(entityarrow.getDamage() + (double)i * 0.5D + 0.5D);
+            entityarrow.setDamage(entityarrow.getDamage() + i * 0.5D + 0.5D);
         }
 
         if (j > 0)
